+++ src/index.ts
@@ -1,2 +1,3 @@
+export { default as cleanTestDb } from "./cleanTestDb.js"
 export { default as provideDreamViteMatchers } from "./viteMatchers.js"
 export { default as truncate } from "./truncate.js"